Enum modulo_traits::file_msg::ToFileThreadMsg

pub enum ToFileThreadMsg {
    ReplaceText(Point, Option<Point>, String),
    ClearAllText,
    Save(Sender<SaveResult>),
}

Messages that can be sent to the file thread to manipulate text

Variants

ReplaceText

Replace text between the start point and end point with the data from the string. If the end point is None, the text is just inserted at the start point. If the string is the empty string, the text between the start point and end point is deleted.

ClearAllText

Clears all the text in the file

Save

Saves the file to the OS file system. If the path is None, the file will not be saved. The front end is responsible for making sure the file thread has a proper path to a file. This may in the future accept a channel that